vhost: introduce vhost transport operations structure
This is the first of a series of patches, whose purpose is to add
support for the virtio-vhost-user transport. This is a vhost-user
transport implementation that is different from the default AF_UNIX
transport. It uses the virtio-vhost-user PCI device in order to tunnel
vhost-user protocol messages over virtio. This lets guests act as vhost
device backends for other guests.

File descriptor passing is specific to the AF_UNIX vhost-user protocol
transport.  In order to add support for additional transports, it is
necessary to extract transport-specific code from the main vhost-user
code.

This patch introduces struct vhost_transport_ops and associates each
device with a transport.  Core vhost-user code calls into
vhost_transport_ops to perform transport-specific operations.

Notifying callfd is a transport-specific operation, so it belongs to
trans_af_unix.c.

Several more patches follow this one to complete the task of moving
AF_UNIX transport code out of core vhost-user code.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>