drivers: add virtio-vhost-user transport
This patch introduces the virtio-vhost-user transport. This transport is
based on the virtio-vhost-user device. This device replaces the AF_UNIX
socket used by the vhost-user protocol with a virtio device that tunnels
vhost-user protocol messages.  This allows a guest to act as a vhost
device backend for other guests.

For more information on virtio-vhost-user, see
https://wiki.qemu.org/Features/VirtioVhostUser.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>

* Change directory structures
* Write meson.build instead of Makefile
* Fix logger functions in trans_virtio_vhost_user.c
* Remove check for VHOST_USER_DEQUEUE_ZERO_COPY in
  trans_virtio_vhost_user.c

Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>